Working towards self-sufficiency.
Currently, we use quite some resources that have to be built outside of
Joyce. The current changes allow to download ontology class mappings
from within Joyce and also to extract class names and synonyms (the
latter has to be done by calling a separate process due to an OWL API
version clash between the julielab bioportal tools and OAPT).